permutation-in-string/main_test.go
package permutation_in_string

import "testing"

func Test_checkInclusion(t *testing.T) {
	type args struct {
		s1 string
		s2 string
	}
	tests := []struct {
		name string
		args args
		want bool
	}{
		{
			name: "ex0",
			args: args{
				s1: "ab",
				s2: "eidbaooo",
			},
			want: true,
		},
		{
			name: "ex1",
			args: args{
				s1: "ab",
				s2: "eidboaooo",
			},
			want: false,
		},
		{
			name: "ex2",
			args: args{
				s1: "ab",
				s2: "ba",
			},
			want: true,
		},
		{
			name: "ex3",
			args: args{
				s1: "adc",
				s2: "dcda",
			},
			want: true,
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := checkInclusion(tt.args.s1, tt.args.s2); got != tt.want {
				t.Errorf("checkInclusion() = %v, want %v", got, tt.want)
			}

			if got := checkInclusionO1(tt.args.s1, tt.args.s2); got != tt.want {
				t.Errorf("checkInclusion() = %v, want %v", got, tt.want)
			}

			if got := checkInclusionO2(tt.args.s1, tt.args.s2); got != tt.want {
				t.Errorf("checkInclusion() = %v, want %v", got, tt.want)
			}

			if got := checkInclusionO3(tt.args.s1, tt.args.s2); got != tt.want {
				t.Errorf("checkInclusion() = %v, want %v", got, tt.want)
			}
		})
	}
}
